refactor: task 관련 type 위치 변경
