+++ package.json
@@ -38,7 +38,8 @@
     "react-redux": "^8.1.3",
     "react-toastify": "^9.1.3",
     "server-only": "^0.0.1",
-    "yup": "^1.3.2"
+    "yup": "^1.3.2",
+    "zustand": "^4.5.0"
   },
   "devDependencies": {
     "@svgr/webpack": "^8.1.0",
