+++ src/components/organisms/taskList/TaskList.hook.tsx
@@ -1,10 +1,8 @@
 import { useClientTranslation } from '@/libs/i18n/useClientTranslation';
 import { useState } from 'react';
 import { useDispatch, useSelector } from '@/libs/redux';
-import taskSlice, {
-  IsCompletedFilter,
-  PeriodFilter,
-} from '@/libs/redux/slices/taskSlice';
+import taskSlice from '@/libs/redux/slices/taskSlice';
+import { IsCompletedFilter, PeriodFilter } from '@/types/task/task.type';
 
 const useTaskList = () => {
   const { t } = useClientTranslation('task');
